+++ examples/wow/zone_features__pandas.py
@@ -3,14 +3,6 @@ import pandas as pd
 from hamilton.function_modifiers import extract_columns
 
 
-def darkshore_flag(zone: pd.Series) -> pd.Series:
-    return pd.Series((zone == "Darkshore").astype(int), index=zone.index)
-
-
-def durotar_flag(zone: pd.Series) -> pd.Series:
-    return pd.Series((zone == "Durotar").astype(int), index=zone.index)
-
-
 def with_flags(
     avatarId: pd.Series, darkshore_flag: pd.Series, durotar_flag: pd.Series
 ) -> pd.DataFrame:
@@ -26,11 +18,3 @@ def zone_counts(with_flags: pd.DataFrame, aggregation_level: str) -> pd.DataFram
         darkshore_count=("darkshore_flag", "sum"),
         durotar_count=("durotar_flag", "sum"),
     )
-
-
-def darkshore_likelihood(darkshore_count: pd.Series, total_count: pd.Series) -> pd.Series:
-    return darkshore_count / total_count
-
-
-def durotar_likelihood(durotar_count: pd.Series, total_count: pd.Series) -> pd.Series:
-    return durotar_count / total_count
