+++ examples/wow/zone_features__spark_v2.py
@@ -1,8 +1,7 @@
-import pandas as pd
 import pyspark.sql as ps
 from pyspark.sql import functions as sf
+from zone_features import darkshore_flag, darkshore_likelihood, durotar_flag, durotar_likelihood
 
-from hamilton import htypes
 from hamilton.plugins.h_spark import with_columns
 
 
@@ -14,19 +13,7 @@ def world_of_warcraft(spark_session: ps.SparkSession) -> ps.DataFrame:
     return spark_session.read.parquet("data/wow.parquet")
 
 
-def _flag_functions():
-    """Hidden from the DAG by using this wrapper function"""
-
-    def durotar_flag(zone: pd.Series) -> htypes.column[pd.Series, int]:
-        return (zone == " Durotar").astype(int)
-
-    def darkshore_flag(zone: pd.Series) -> htypes.column[pd.Series, int]:
-        return (zone == " Darkshore").astype(int)
-
-    return durotar_flag, darkshore_flag
-
-
-@with_columns(*_flag_functions(), columns_to_pass=["zone"])
+@with_columns(darkshore_flag, durotar_flag, columns_to_pass=["zone"])
 def with_flags(world_of_warcraft: ps.DataFrame, aggregation_level: str) -> ps.DataFrame:
     return world_of_warcraft
 
@@ -39,22 +26,10 @@ def zone_counts(with_flags: ps.DataFrame, aggregation_level: str) -> ps.DataFram
     )
 
 
-def _likelihood_functions():
-    def durotar_likelihood(
-        durotar_count: pd.Series, total_count: pd.Series
-    ) -> htypes.column[pd.Series, float]:
-        return durotar_count / total_count
-
-    def darkshore_likelihood(
-        darkshore_count: pd.Series, total_count: pd.Series
-    ) -> htypes.column[pd.Series, float]:
-        return darkshore_count / total_count
-
-    return durotar_likelihood, darkshore_likelihood
-
-
 @with_columns(
-    *_likelihood_functions(), columns_to_pass=["durotar_count", "darkshore_count", "total_count"]
+    darkshore_likelihood,
+    durotar_likelihood,
+    columns_to_pass=["durotar_count", "darkshore_count", "total_count"],
 )
 def zone_likelihoods(zone_counts: ps.DataFrame) -> ps.DataFrame:
     return zone_counts
